Maybe fix river test

+++ river/river_test.go
@@ -28,7 +28,7 @@ func TestRiver(t *testing.T) {
 	assert.WithinDuration(time.Now(), v.Metadata.WhenGMT.Time, time.Second)
 	assert.WithinDuration(time.Now(), v.Metadata.WhenLocal.Time, time.Second)
 	assert.Equal("3", v.Metadata.Version)
-	assert.Equal(0, v.Metadata.Secs)
+	assert.Equal(float64(0), v.Metadata.Secs)
 
 	assert.Equal(models.Feeds{[]models.Feed{}}, v.UpdatedFeeds)
 }
